ATCubicSplineFit added to ATTrajectory ATCubicSplineFit is now calculated on smoothed cloud
